Update tooltip to use Radix UI instead of Reach UI
