+++ app/client/src/components/app.tsx
@@ -8,7 +8,7 @@ import {
   useLocation,
 } from "react-router-dom";
 import { useIdleTimer } from "react-idle-timer";
-import "@reach/tooltip/styles.css";
+import { TooltipProvider } from "@radix-ui/react-tooltip";
 import "uswds/css/uswds.css";
 import "uswds/js/uswds.js";
 import "bootstrap/dist/css/bootstrap-grid.min.css";
@@ -269,10 +269,10 @@ function ProtectedRoute({ children }: { children: JSX.Element }) {
   }
 
   return (
-    <>
+    <TooltipProvider>
       <ConfirmationDialog />
       {children}
-    </>
+    </TooltipProvider>
   );
 }
